backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
pydantic==2.5.0 # SQLModel 0.0.14 is compatible with Pydantic v2
# httpx==0.25.2 # Removed, consolidated below
python-dotenv==1.0.0

# LLM SDKs
openai~=1.23.0
anthropic~=0.21.0
google-generativeai~=0.7.0

# Testing
pytest~=8.0.0 # Using a recent version of pytest
httpx~=0.27.0 # httpx is already a dependency of fastapi, but good to specify for TestClient

# sqlalchemy is a dependency of sqlmodel
# sqlite3 is a built-in module, not needed here
